refactor: dqn agents
