+++ turtlebot3_dqn/nodes/dqn_agent.py
@@ -16,25 +16,25 @@
 #################################################################################
 
 # Authors: Gilbert, Widowski, Mueller #
-
 import rospy
 import os
-import json
-import numpy as np
-import random
 import sys
+import json
+import random
+import numpy as np
 
 sys.path.append(os.path.dirname(os.path.abspath(os.path.dirname(__file__))))
-from hindsight_experience_replay import HindsightExperienceReplay
+
+from collections import deque
 from std_msgs.msg import Float32MultiArray
 from keras.models import Sequential, load_model
 from keras.optimizers import RMSprop
-from keras.layers import Dense, Dropout, Activation
+from keras.layers import Dense, Dropout, Activation, BatchNormalization
 
 
 class ReinforceAgent:
-    def __init__(self, state_size, action_size, goal_size, stage="1",
-                 episode_max_steps=6000, target_update=2000, discount_factor=0.99,
+    def __init__(self, state_size, action_size, stage="1",
+                 episode_max_step=6000, target_update=2000, discount_factor=0.99,
                  learning_rate=0.00025, epsilon=1.0, epsilon_decay=0.99,
                  epsilon_min=0.05, batch_size=64, train_start=64, load_model_bool=True):
         self.pub_result = rospy.Publisher('result', Float32MultiArray, queue_size=5)
@@ -42,11 +42,11 @@ class ReinforceAgent:
         self.dirPath = self.dirPath.replace('turtlebot3_dqn/nodes', 'turtlebot3_dqn/save_model/stage_' + stage + '_')
         self.result = Float32MultiArray()
 
+        self.load_model = load_model_bool
         self.load_episode = "latest"
         self.state_size = state_size
         self.action_size = action_size
-        self.goal_size = goal_size
-        self.episode_max_steps = episode_max_steps
+        self.episode_max_step = episode_max_step
         self.target_update = target_update
         self.discount_factor = discount_factor
         self.learning_rate = learning_rate
@@ -56,13 +56,12 @@ class ReinforceAgent:
         self.batch_size = batch_size
         self.train_start = train_start
 
-        self.her = HindsightExperienceReplay(k=1, strategie="future", maxlen=1000000, batch_size=self.batch_size)
-        self.q_values = np.zeros(self.action_size)
-
+        self.memory = deque(maxlen=1000000)
+        self.q_value = np.zeros(self.action_size)
         self.model = self.build_model()
         self.target_model = self.build_model()
-        self.update_target_model()
 
+        self.update_target_model()
         if load_model_bool:
             self.load_saved_model()
 
@@ -85,14 +84,13 @@ class ReinforceAgent:
 
     def build_model(self):
         model = Sequential([
-            Dense(64, input_shape=(self.state_size + self.goal_size), kernel_initializer='lecun_uniform'),
+            Dense(64, input_shape=(self.state_size,), kernel_initializer='lecun_uniform'),
             Activation('relu'),
             Dense(64, kernel_initializer='lecun_uniform'),
             Activation('relu'),
             Dropout(0.2),
-            Dense(self.action_size, kernel_initializer='lecun_uniform')
+            Dense(self.action_size, kernel_initializer='lecun_uniform'),
         ])
-
         model.compile(loss='mse', optimizer=RMSprop(lr=self.learning_rate, rho=0.9, epsilon=1e-06))
         model.summary()
 
@@ -107,48 +105,49 @@ class ReinforceAgent:
     def update_target_model(self):
         self.target_model.set_weights(self.model.get_weights())
 
-    def predict(self, state, goal):
-        features = np.hstack((np.asarray(state).flatten(), np.asarray(goal).flatten()))
-        return self.model.predict(features.reshape(1, len(state) + len(goal)))
-
-    def predict_target(self, state, goal):
-        features = np.hstack((np.asarray(state).flatten(), np.asarray(goal).flatten()))
-        return self.target_model.predict(features.reshape(1, len(state) + len(goal)))
-
-    def get_action(self, state, goal):
+    def get_action(self, state):
         if np.random.rand() <= self.epsilon:
-            self.q_values = np.zeros(self.action_size)
+            self.q_value = np.zeros(self.action_size)
             return random.randrange(self.action_size)
         else:
-            q_value = self.predict(state, goal)
-            self.q_values = q_value
+            q_value = self.model.predict(state.reshape(1, len(state)))
+            self.q_value = q_value
             return np.argmax(q_value[0])
 
+    def append_memory(self, state, action, reward, next_state, done):
+        self.memory.append((state, action, reward, next_state, done))
+
     def train_model(self, target=False):
-        mini_batch = self.her.sample_memory()
-        X_batch = np.empty((0, self.state_size + self.goal_size), dtype=np.float64)
+        mini_batch = random.sample(self.memory, self.batch_size)
+        X_batch = np.empty((0, self.state_size), dtype=np.float64)
         Y_batch = np.empty((0, self.action_size), dtype=np.float64)
 
         for i in range(self.batch_size):
-            states, actions, goals, rewards, next_states, dones = mini_batch[i]
-            self.q_values = self.predict(states, goals)
+            states = mini_batch[i][0]
+            actions = mini_batch[i][1]
+            rewards = mini_batch[i][2]
+            next_states = mini_batch[i][3]
+            dones = mini_batch[i][4]
+
+            q_value = self.model.predict(states.reshape(1, len(states)))
+            self.q_value = q_value
 
             if target:
-                next_target = self.predict_target(next_states, goals)
+                next_target = self.target_model.predict(next_states.reshape(1, len(next_states)))
 
             else:
-                next_target = self.predict(next_states, goals)
+                next_target = self.model.predict(next_states.reshape(1, len(next_states)))
 
-            next_q_values = self.get_q_value(rewards, next_target, dones)
+            next_q_value = self.get_q_value(rewards, next_target, dones)
 
-            X_batch = np.append(X_batch, np.asarray([np.hstack((states, goals)).copy()]), axis=0)
-            Y_sample = self.q_values.copy()
+            X_batch = np.append(X_batch, np.array([states.copy()]), axis=0)
+            Y_sample = q_value.copy()
 
-            Y_sample[0][actions] = next_q_values
+            Y_sample[0][actions] = next_q_value
             Y_batch = np.append(Y_batch, np.array([Y_sample[0]]), axis=0)
 
             if dones:
-                X_batch = np.append(X_batch, np.asarray([np.hstack((next_states, goals)).copy()]), axis=0)
+                X_batch = np.append(X_batch, np.array([next_states.copy()]), axis=0)
                 Y_batch = np.append(Y_batch, np.array([[rewards] * self.action_size]), axis=0)
 
         self.model.fit(X_batch, Y_batch, batch_size=self.batch_size, epochs=1, verbose=0)
